fix(ffx-api): move ids into a single file to avoid circular dependency
