+++ packages/ffx-api/src/index.mts
@@ -1,6 +1,5 @@
 import {
   Agent,
-  AgentId,
   Agents,
   CreateAgentInput,
   createAgent,
@@ -11,7 +10,6 @@ import {
 import {
   CreateDocumentInput,
   Document,
-  DocumentId,
   Documents,
   UpdateDocumentInput,
   createDocument,
@@ -23,7 +21,6 @@ import {
 import {
   CreateEnvironmentInput,
   Environment,
-  EnvironmentId,
   Environments,
   UpdateEnvironmentInput,
   createEnvironment,
@@ -33,19 +30,27 @@ import {
   updateEnvironment,
 } from "./lib/environments.mjs";
 import {
+  AgentId,
+  DocumentId,
+  EnvironmentId,
+  SecretId,
+  SheetId,
+  SpaceId,
+  VersionId,
+  WorkbookId,
+} from "./lib/ids.mjs";
+import {
   CreateSecretInput,
   Secret,
-  SecretId,
   Secrets,
   createSecret,
   deleteSecret,
   listSecrets,
 } from "./lib/secrets.mjs";
-import { Sheet, SheetId, Sheets, deleteSheet, getSheet, listSheets } from "./lib/sheets.mjs";
+import { Sheet, Sheets, deleteSheet, getSheet, listSheets } from "./lib/sheets.mjs";
 import {
   CreateSpaceInput,
   Space,
-  SpaceId,
   Spaces,
   UpdateSpaceInput,
   archiveSpace,
@@ -56,12 +61,11 @@ import {
   updateSpace,
 } from "./lib/spaces.mjs";
 import { ApiReader, DecoderErrors, HttpError, Successful } from "./lib/types.mjs";
-import { Version, VersionId, createVersion } from "./lib/versions.mjs";
+import { Version, createVersion } from "./lib/versions.mjs";
 import {
   CreateWorkbookInput,
   UpdateWorkbookInput,
   Workbook,
-  WorkbookId,
   Workbooks,
   createWorkbook,
   deleteWorkbook,
@@ -233,19 +237,31 @@ export default function mkApiClient(secret: string, environmentId: EnvironmentId
   };
 }
 
-export { Agent, AgentId, Agents, EventTopic, isoAgentId } from "./lib/agents.mjs";
-export { Document, DocumentId, Documents, isoDocumentId } from "./lib/documents.mjs";
+export { Agent, Agents, EventTopic } from "./lib/agents.mjs";
+export { Document, Documents } from "./lib/documents.mjs";
+export { Environment, Environments } from "./lib/environments.mjs";
 export {
-  AccountId,
-  Environment,
-  Environments,
+  AgentId,
+  DocumentId,
   EnvironmentId,
+  SecretId,
+  SheetId,
+  SpaceId,
+  VersionId,
+  WorkbookId,
   isoAccountId,
+  isoAgentId,
+  isoDocumentId,
   isoEnvironmentId,
-} from "./lib/environments.mjs";
-export { Secret, SecretId, Secrets, isoSecretId } from "./lib/secrets.mjs";
-export { Permission, Sheet, SheetId, Sheets, isoSheetId } from "./lib/sheets.mjs";
-export { Space, SpaceId, Spaces, isoSpaceId } from "./lib/spaces.mjs";
-export { UserId, isoUserId } from "./lib/users.mjs";
-export { Version, VersionId, isoVersionId } from "./lib/versions.mjs";
-export { Workbook, WorkbookId, Workbooks, isoWorkbookId } from "./lib/workbooks.mjs";
+  isoSecretId,
+  isoSheetId,
+  isoSpaceId,
+  isoUserId,
+  isoVersionId,
+  isoWorkbookId,
+} from "./lib/ids.mjs";
+export { Secret, Secrets } from "./lib/secrets.mjs";
+export { Permission, Sheet, Sheets } from "./lib/sheets.mjs";
+export { Space, Spaces } from "./lib/spaces.mjs";
+export { Version } from "./lib/versions.mjs";
+export { Workbook, Workbooks } from "./lib/workbooks.mjs";
