+++ packages/ffx-api/src/lib/documents.mts
@@ -2,14 +2,16 @@ import axios, { AxiosError } from "axios";
 import { identity, pipe } from "fp-ts/function";
 import * as RT from "fp-ts/ReaderTask";
 import * as RTE from "fp-ts/ReaderTaskEither";
-import * as Str from "fp-ts/string";
 import * as TE from "fp-ts/TaskEither";
 import * as t from "io-ts";
-import { Iso } from "monocle-ts";
-import { Newtype, iso } from "newtype-ts";
 
-import { EnvironmentIdFromString } from "./environments.mjs";
-import { SpaceId, SpaceIdFromString } from "./spaces.mjs";
+import {
+  DocumentId,
+  DocumentIdFromString,
+  EnvironmentIdFromString,
+  SpaceId,
+  SpaceIdFromString,
+} from "./ids.mjs";
 import {
   ApiReader,
   DecoderErrors,
@@ -22,23 +24,6 @@ import {
 // ==================
 //   Runtime codecs
 // ==================
-
-export interface DocumentId extends Newtype<{ readonly DocumentId: unique symbol }, string> {}
-
-export const isoDocumentId: Iso<DocumentId, string> = iso<DocumentId>();
-
-export const DocumentIdFromString = new t.Type<DocumentId>(
-  "DocumentIdFromString",
-  (input: unknown): input is DocumentId => {
-    return Str.isString(input) && /^us_dc_\w{8}$/g.test(input);
-  },
-  (input, context) => {
-    return Str.isString(input) && /^us_dc_\w{8}$/g.test(input)
-      ? t.success(isoDocumentId.wrap(input))
-      : t.failure(input, context);
-  },
-  t.identity,
-);
 
 export const DocumentC = t.intersection(
   [
