+++ packages/ffx-api/src/lib/environments.mts
@@ -2,12 +2,10 @@ import axios, { AxiosError } from "axios";
 import { identity, pipe } from "fp-ts/function";
 import * as RT from "fp-ts/ReaderTask";
 import * as RTE from "fp-ts/ReaderTaskEither";
-import * as Str from "fp-ts/string";
 import * as TE from "fp-ts/TaskEither";
 import * as t from "io-ts";
-import { Iso } from "monocle-ts";
-import { Newtype, iso } from "newtype-ts";
 
+import { AccountIdFromString, EnvironmentId, EnvironmentIdFromString } from "./ids.mjs";
 import {
   ApiReader,
   DecoderErrors,
@@ -22,40 +20,6 @@ import {
 // ==================
 
 export const AuthLinkC = t.union([t.literal("shared_link"), t.literal("magic_link")]);
-
-export interface EnvironmentId extends Newtype<{ readonly EnvironmentId: unique symbol }, string> {}
-
-export const isoEnvironmentId: Iso<EnvironmentId, string> = iso<EnvironmentId>();
-
-export const EnvironmentIdFromString = new t.Type<EnvironmentId>(
-  "EnvironmentIdFromString",
-  (input: unknown): input is EnvironmentId => {
-    return Str.isString(input) && /^us_env_\w{8}$/g.test(input);
-  },
-  (input, context) => {
-    return Str.isString(input) && /^us_env_\w{8}$/g.test(input)
-      ? t.success(isoEnvironmentId.wrap(input))
-      : t.failure(input, context);
-  },
-  t.identity,
-);
-
-export interface AccountId extends Newtype<{ readonly AccountId: unique symbol }, string> {}
-
-export const isoAccountId: Iso<AccountId, string> = iso<AccountId>();
-
-export const AccountIdFromString = new t.Type<AccountId>(
-  "AccountIdFromString",
-  (input: unknown): input is AccountId => {
-    return Str.isString(input) && /^us_acc_\w{8}$/g.test(input);
-  },
-  (input, context) => {
-    return Str.isString(input) && /^us_acc_\w{8}$/g.test(input)
-      ? t.success(isoAccountId.wrap(input))
-      : t.failure(input, context);
-  },
-  t.identity,
-);
 
 export const EnvironmentC = t.intersection(
   [
