+++ packages/ffx-api/src/lib/secrets.mts
@@ -2,14 +2,17 @@ import axios, { AxiosError } from "axios";
 import { identity, pipe } from "fp-ts/function";
 import * as RT from "fp-ts/ReaderTask";
 import * as RTE from "fp-ts/ReaderTaskEither";
-import * as Str from "fp-ts/string";
 import * as TE from "fp-ts/TaskEither";
 import * as t from "io-ts";
-import { Iso } from "monocle-ts";
-import { Newtype, iso } from "newtype-ts";
 
-import { EnvironmentId, EnvironmentIdFromString } from "./environments.mjs";
-import { SpaceId, SpaceIdFromString } from "./spaces.mjs";
+import {
+  EnvironmentId,
+  EnvironmentIdFromString,
+  SecretId,
+  SecretIdFromString,
+  SpaceId,
+  SpaceIdFromString,
+} from "./ids.mjs";
 import {
   ApiReader,
   DecoderErrors,
@@ -22,23 +25,6 @@ import {
 // ==================
 //   Runtime codecs
 // ==================
-
-export interface SecretId extends Newtype<{ readonly SecretId: unique symbol }, string> {}
-
-export const isoSecretId: Iso<SecretId, string> = iso<SecretId>();
-
-export const SecretIdFromString = new t.Type<SecretId>(
-  "SecretIdFromString",
-  (input: unknown): input is SecretId => {
-    return Str.isString(input) && /^us_sec_\w{8}$/g.test(input);
-  },
-  (input, context) => {
-    return Str.isString(input) && /^us_sec_\w{8}$/g.test(input)
-      ? t.success(isoSecretId.wrap(input))
-      : t.failure(input, context);
-  },
-  t.identity,
-);
 
 export const SecretC = t.intersection(
   [
