+++ packages/ffx-api/test/agents.spec.mts
@@ -7,8 +7,8 @@ import { setupServer } from "msw/node";
 import { match } from "ts-pattern";
 
 import mkApiClient from "../src/index.mjs";
-import { Agent, AgentC, AgentIdFromString, Agents, isoAgentId } from "../src/lib/agents.mjs";
-import { EnvironmentId, isoEnvironmentId } from "../src/lib/environments.mjs";
+import { Agent, AgentC, Agents } from "../src/lib/agents.mjs";
+import { AgentIdFromString, EnvironmentId, isoAgentId, isoEnvironmentId } from "../src/lib/ids.mjs";
 
 function randomId(): IO.IO<string> {
   return IO.of(Math.random().toString(16).slice(2, 10));
