+++ packages/ffx-api/test/documents.spec.mts
@@ -7,9 +7,15 @@ import { setupServer } from "msw/node";
 import { match } from "ts-pattern";
 
 import mkApiClient from "../src/index.mjs";
-import { Document, DocumentC, DocumentIdFromString, isoDocumentId } from "../src/lib/documents.mjs";
-import { EnvironmentId, isoEnvironmentId } from "../src/lib/environments.mjs";
-import { SpaceIdFromString, isoSpaceId } from "../src/lib/spaces.mjs";
+import { Document, DocumentC } from "../src/lib/documents.mjs";
+import {
+  DocumentIdFromString,
+  EnvironmentId,
+  isoDocumentId,
+  isoEnvironmentId,
+  isoSpaceId,
+  SpaceIdFromString,
+} from "../src/lib/ids.mjs";
 
 function randomId(): IO.IO<string> {
   return IO.of(Math.random().toString(16).slice(2, 10));
