+++ packages/ffx-api/test/environments.spec.mts
@@ -7,15 +7,14 @@ import { setupServer } from "msw/node";
 import { match } from "ts-pattern";
 
 import mkApiClient from "../src/index.mjs";
+import { Environment, EnvironmentC } from "../src/lib/environments.mjs";
 import {
   AccountIdFromString,
-  Environment,
-  EnvironmentC,
   EnvironmentId,
   EnvironmentIdFromString,
   isoAccountId,
   isoEnvironmentId,
-} from "../src/lib/environments.mjs";
+} from "../src/lib/ids.mjs";
 
 function randomId(): IO.IO<string> {
   return IO.of(Math.random().toString(16).slice(2, 10));
