+++ packages/ffx-api/test/versions.spec.mts
@@ -7,15 +7,16 @@ import { setupServer } from "msw/node";
 import { match } from "ts-pattern";
 
 import mkApiClient from "../src/index.mjs";
-import { EnvironmentId, isoEnvironmentId } from "../src/lib/environments.mjs";
-import { SheetId, isoSheetId } from "../src/lib/sheets.mjs";
 import {
-  Version,
-  VersionC,
+  EnvironmentId,
+  SheetId,
   VersionId,
   VersionIdFromString,
+  isoEnvironmentId,
+  isoSheetId,
   isoVersionId,
-} from "../src/lib/versions.mjs";
+} from "../src/lib/ids.mjs";
+import { Version, VersionC } from "../src/lib/versions.mjs";
 
 function randomId(): IO.IO<string> {
   return IO.of(Math.random().toString(16).slice(2, 10));
